refactor(backend): extract controller logic from app.js to tracks controller
