+++ backend/src/app.js
@@ -1,32 +1,14 @@
 const express = require("express");
-const { Authenticate, GetTracks, TreatUntreatedTracks } = require("./services");
+const { Tracks } = require("./controllers");
 
 const app = express();
 const PORT = process.env.PORT || 3000;
 
-const ISRCs = [
-  "US7VG1846811",
-  "US7QQ1846811",
-  "BRC310600002",
-  "BR1SP1200071",
-  "BR1SP1200070",
-  "BR1SP1500002",
-  "BXKZM1900338",
-  "BXKZM1900345",
-  "QZNJX2081700",
-  "QZNJX2078148"
-];
-
 app.get('/tracks', async (_, res) => {
-  const authenticateService = new Authenticate();
-  const treatUntreatedTracksService = new TreatUntreatedTracks();
-  const getTracksService = new GetTracks(
-    authenticateService,
-    treatUntreatedTracksService
-  );
+  const tracksController = new Tracks(); 
   
   try {
-    const tracks = await getTracksService.getTracksByISRCs(ISRCs);
+    const tracks = await tracksController.getTracks();
     res.status(200).json({ data: tracks });
   } catch (error) {
     console.log(error);
