+++ backend/src/controllers/index.js
@@ -1,0 +1,5 @@
+const { Tracks } = require("./Tracks");
+
+module.exports = {
+  Tracks
+};
